requirements.txt
uWSGI
falcon
PyYAML
easydict
psycopg2
